code/requirements.txt
google-cloud-aiplatform>=1.63.0,<2.0.0
langchain-google-vertexai>=1.0.10,<2.0.0
langchain>=0.2.14,<0.3.0
langchain-community>=0.2.14,<0.3.0
beautifulsoup4>=4.12.3,<5.0.0
python-dotenv>=1.0.1,<2.0.0
langchain-text-splitters>=0.2.2,<0.3.0
tiktoken>=0.7.0,<0.8.0
atlassian-python-api>=3.41.14,<4.0.0
google-cloud>=0.34.0,<1.0.0
uvicorn>=0.30.6,<0.31.0
fastapi>=0.115.0,<0.116.0
